[fixes #26] Rename LMS & NLMS filter classes
