+++ src/test/java/com/github/psambit9791/jdsp/TestLMS.java
@@ -1,6 +1,6 @@
 package com.github.psambit9791.jdsp;
 
-import com.github.psambit9791.jdsp.filter.adaptive.NLMSFilter;
+import com.github.psambit9791.jdsp.filter.adaptive.LMS;
 import com.github.psambit9791.jdsp.misc.Plotting;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Order;
@@ -10,7 +10,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
 
-public class TestNLMSFilter {
+public class TestLMS {
     // Random signal, with noise added
     private final double[] signal1 = {7.73615076e-01, -2.38644082e-01, -1.44153846e-02, -7.43728154e-02, 1.28832364e+00, -8.86944508e+00,  4.11744669e+00,  2.78178586e+00, -5.03828060e-01,  3.97897524e+00,  5.56136074e+00,  4.33681482e+00, 5.26247771e-01, -8.12778645e-01,  2.51919943e+00,  1.00856762e+00, 7.06685871e+00, -4.66127298e+00,  2.03547177e-01, -5.25186465e-01, -4.87098863e+00, -1.63898740e+00,  1.87870253e-01, -5.57965145e-01, -1.07066598e-01,  9.19816384e+00, -8.71136174e-01, -3.02327798e+00, -5.82031931e-02, -1.16211219e-01,  1.13151235e-01,  1.26114690e+00, 4.73462459e+00,  1.53611870e+00, -9.98205847e-01, -3.28775428e+00, 9.07053454e+00, -1.54833550e+00, -1.51008254e+00,  6.51624328e+00, -4.55701930e+00,  8.62041101e-01, -1.11631762e+00,  6.12632729e+00, 1.55095287e-01,  6.17289083e-01,  5.53387303e+00,  3.98012854e+00, -2.37425479e+00, -6.35874830e+00, -4.83330811e+00,  8.46856491e-01, -2.93207237e+00,  5.09877419e+00, -9.05449003e+00,  1.48636457e-01, -3.58520346e+00,  4.23115938e+00,  3.97121083e+00,  1.41538968e+00, -9.31974362e-01, -4.85514328e+00, -4.14552444e+00, -2.16680695e+00, 5.04741213e+00, -6.94276453e-02, -5.90503480e+00, -3.59451409e+00, 6.36499159e+00,  1.01308575e+01,  3.41732387e+00,  5.42015787e+00, 3.78165763e+00,  5.61407630e+00,  2.03656936e+00, -2.31467356e+00, 5.95036932e+00,  2.44526710e+00, -7.40024361e-01,  7.98213484e-02, -3.06403105e+00, -5.86909930e-01, -4.89830029e+00,  1.23679666e+01, 9.91312683e-02,  4.23599196e+00,  6.27245291e-01,  3.52968840e+00, 5.00869852e+00, -1.50513410e+00, -2.16011962e+00,  5.49917811e-01, -2.26659683e+00,  1.70725513e+00,  2.39727075e+00,  5.15351990e+00, 3.11771504e+00,  9.32193978e-01, -8.19075713e-02, -3.61008086e+00, -4.19740389e+00, -6.73014522e+00,  5.73541558e+00, -7.23020975e+00, -4.46432861e+00,  1.03575059e+01, -5.39284152e+00,  4.72954562e-02, 5.80744105e+00, -1.51942844e+00, -4.65530480e-01, -5.71813655e-01, 3.05234053e+00,  6.09075515e-01, -1.98037506e+00,  7.14506231e-01, 2.68107081e+00,  3.24380756e+00, -7.08060738e+00, -1.71634391e+00, 5.06864995e+00, -4.31711618e+00,  3.87541230e+00, -8.49148767e+00, 2.66471079e+00, -3.74045465e+00,  4.99362609e+00, -5.12173211e+00, -1.33151021e-01,  8.36393473e-01,  5.74386790e+00, -2.64544210e+00, 4.50639718e+00,  7.28226783e+00,  2.15130604e-01,  2.03331264e+00, -6.40605600e+00,  1.69201321e+00,  5.21920279e+00,  2.84600780e-01, 2.34699360e+00, -2.54902107e+00, -2.66004164e+00, -2.39064700e+00, -6.32328222e-01,  9.82091206e-01, -6.51881138e-02, -3.04311839e+00, -1.76498839e+00,  2.66959255e+00,  8.37398729e-01, -3.25932388e+00, 6.51018440e+00,  2.02981123e+00, -5.41678580e+00,  2.06063144e+00, -7.70799133e-01,  2.57805433e+00,  2.91215704e+00, -4.41207280e+00, -4.67975016e+00, -6.61825886e+00, -3.35400589e+00,  8.74674627e-03, 7.43617310e-01, -3.75279695e+00, -3.27974734e-01, -2.73018881e+00, -4.66419127e+00, -7.27015174e+00, -3.65824249e+00, -1.08689861e+01, -6.81145397e-01, -5.06373799e-01, -1.02286408e+00,  6.91061508e-01, -3.08047210e+00,  5.71097821e+00, -6.78346454e+00, -6.01614922e+00, -1.75321783e+00,  1.06042567e+00,  3.29761152e+00, -1.22316856e+01, 9.22934835e-01, -9.55571638e+00, -1.15720607e+00, -7.53973142e+00, 1.53614251e+00,  3.27826037e-01, -3.75613319e+00,  1.74610563e+00, 5.68450799e+00, -1.25439441e+00,  6.34719872e+00, -8.76148347e-02, -9.36224947e-01,  2.82739862e-01,  1.32286366e+00,  1.32513382e+00, 2.93519406e+00, -8.29501002e-02, -9.35219340e-01,  9.52288323e-01, 2.25459000e+00,  3.03397928e+00, -1.50991261e+00,  1.89007422e+00, 2.81578034e-01, -3.45521704e+00, -2.38102363e+00, -2.49901416e+00, 8.25811154e+00,  3.08093564e+00, -1.17275450e+00,  2.29109765e+00, 9.66588605e-01, -7.29614680e+00, -3.89862770e+00, -1.63177776e+00, 2.21062239e+00, -5.62142151e+00,  3.28703998e+00,  8.52994288e+00, -1.24469719e+01, -4.36683316e+00, -2.72106177e-02,  4.62359183e+00, 7.24635389e+00, -1.62475247e+00, -7.66715584e+00,  1.30017340e+00, -5.62427464e+00,  1.30661338e+00,  1.05783786e+00,  3.16697154e-01, -5.08342315e-01,  4.91503104e+00,  7.68950904e-01, -8.37960473e+00, -1.40923254e+00, -2.63983294e-01,  4.57667095e+00, -7.61954565e+00, -5.15705438e-01,  2.68324181e+00,  1.53501494e-01,  1.06547807e+00, -2.52481266e+00,  4.02400818e+00,  2.94278906e+00, -2.59587851e+00, 4.76356947e+00,  3.18790104e+00, -5.30617141e-01,  3.22259577e+00, 6.98506914e+00,  5.26638339e+00, -2.96500709e+00, -4.51119140e-01, 1.49355598e+00,  8.66748031e-01,  2.41373830e+00, -1.27962877e+01, -1.69749663e+00, -6.38403922e-01, -2.69025479e-02,  4.02110331e+00, -1.47302099e-01,  4.71460676e+00, -3.71553041e+00,  2.86150123e+00, 4.76119809e+00, -1.19943660e+01,  3.95721120e-01, -8.92553787e+00, -1.97728620e+00,  3.46256920e+00, -7.96185817e+00, -5.68217283e+00, 5.32929913e+00, -1.87257201e+00,  3.82966038e+00,  2.07931439e+00, -5.10278129e+00, -9.75036532e+00,  2.11412442e+00,  1.79136112e+00, 2.08851082e-01, -7.22878268e+00, -4.17436136e+00, -4.63950203e+00, -8.54334383e-01, -5.79718349e+00,  3.86479334e+00,  4.83305034e+00, -4.27911860e+00,  3.54769021e+00, -5.44921171e+00,  8.32057892e+00, -8.00431086e+00, -6.32228896e+00,  7.09907312e-01,  2.33025030e+00, 6.08993453e+00, -6.83040530e+00, -1.64226208e+00, -5.39967380e+00, 3.61227919e+00,  3.59240732e+00,  2.67544663e+00, -2.19368188e+00, 2.49938777e+00,  2.88113362e+00, -3.30686145e+00,  2.53345663e+00, 8.49847765e+00,  6.50954944e+00,  1.73708642e+00,  3.76887276e+00, -3.83969940e+00, -3.99288424e+00, -2.08440684e+00, -6.48882011e+00, -1.17096160e+00, -4.28469410e+00, -6.77145223e-01, -1.71511444e+00, 2.89530722e+00, -2.03556739e+00,  8.01652216e-01, -6.86940911e+00, -1.15869813e+00,  2.26229442e+00,  1.73180528e+00, -3.82690821e+00, -1.30215436e+00,  6.71157000e+00, -5.18619591e+00,  6.10730721e+00, 4.17824364e+00, -5.47350768e-01, -1.95091027e+00,  3.37065184e+00, -1.64087067e+00, -7.07230886e-01,  2.19743101e+00, -2.85092372e+00, 9.61000649e-02, -3.21997942e+00,  5.58314893e+00, -4.74336051e+00, 4.62330740e+00,  1.86626490e-01,  1.37941538e+01, -5.78476826e+00, -5.42956074e-01, -6.20485273e+00, -3.11219820e-01,  4.72771884e-01, -3.00272519e+00,  1.35757357e+01,  3.23723082e+00,  3.52392645e-01, 2.84983460e+00,  5.90285504e+00,  2.78295199e-01, -3.71991593e+00, 4.90393786e-01, -3.23339756e+00,  5.94712326e+00, -3.55754521e+00, 8.96638762e+00,  8.20815729e+00,  3.75698735e+00, -1.63676800e+00, -2.03287800e+00,  6.06536972e+00, -6.24905516e+00,  3.01720879e-02, 4.21378914e+00,  1.38184734e+00,  3.46479778e+00, -1.84767676e+00, -1.03990843e+01, -2.59162399e+00, -1.01643907e-01, -1.91403216e+00, 1.26809915e+00,  2.55410086e+00,  8.14394841e+00, -6.79178072e-02, 5.70672935e+00, -1.04381832e+01,  4.10828311e+00, -1.98631240e+00, 8.05464356e-01,  6.35907464e+00, -2.41284350e+00,  3.54715687e+00, 4.45992563e+00, -8.13578331e+00, -3.49508545e+00, -1.26504169e+00, -3.15447360e+00,  2.40131439e+00,  9.79161803e-01, -3.21828895e+00, 2.15796242e+00, -2.22399922e-02,  3.53969040e-01,  7.00018680e+00, -3.53066462e+00, -4.20977579e-01, -3.23784048e+00,  6.33446347e+00, -2.53304544e-01,  4.36525926e+00, -4.43411558e+00, -1.03541057e+01, 1.00582018e+00,  7.53986173e-01,  9.42740127e+00,  4.23075216e+00, -1.11933540e+00, -6.88447986e+00, -2.22444661e+00, -6.25773066e+00, 6.65462151e-01, -9.94164109e-01, -8.53134292e+00, -3.82421963e+00, 5.47933113e+00,  5.49597131e+00, -1.24903615e+00, -6.18754151e+00, 8.61059760e+00,  4.27719935e+00, -1.65472577e+00, -6.24315843e+00, 7.58833751e+00, -3.19055650e+00, -2.81457455e+00,  1.36777231e-01, 5.74597881e+00,  6.71414287e+00,  9.17304318e+00,  1.59196057e+00, 7.69328546e-01,  2.09118998e+00, -3.38710016e+00,  5.67221909e+00, 5.25098029e+00,  4.19042906e+00,  9.96586250e+00, -5.02617965e+00, -2.05694038e+00,  2.94181580e+00,  3.45276366e-01,  1.00680553e+00, 4.88756801e+00, -2.30546149e+00,  4.76144421e-02,  5.80800680e-01, 4.95026817e+00,  4.53184576e-01, -1.02612348e+00, -3.19471138e+00, -3.08277622e+00, -3.62541739e+00, -1.74932307e-01,  5.64901079e+00, 7.07411639e+00, -1.39018554e+00,  2.46724818e+00, -1.14596684e-01, -2.83008440e-01, -7.08167407e+00, -5.74811827e+00,  1.05915151e+00, 3.27010644e+00,  1.50217680e+00, -5.65604533e-01, -2.86155375e+00, -4.86212742e+00, -1.36804370e-01,  1.19991714e+00,  1.53118287e+00, -1.76843226e+00, -1.98751632e+00, -6.07251008e-02,  2.53082947e+00, -5.37658128e+00,  1.39249309e+00,  4.27613842e+00, -1.35429911e+00, -3.03972901e+00, -8.25716860e+00,  1.85549231e+00,  6.73101071e+00};
     // Random signal, without noise
@@ -32,24 +32,24 @@ public class TestNLMSFilter {
     }
 
     @Test
-    public void NLMSTest1() throws IOException {
-        double mu = 0.1;
+    public void LMSTest1() throws IOException {
+        double mu = 0.01;
         int length = 4;
-        String outputFileName = "test_outputs/NLMSTest1.png";
+        String outputFileName = "test_outputs/LMSTest1.png";
         double[] weights = new double[length];
         Arrays.fill(weights, 0);
 
         // Results calculated with MATLAB R2020b Update 4 9.9.0.1570001
-        double[] result = {0,-1.913567e-02,9.644233e-03,-5.822620e-03,2.319570e-01,-2.280032e+00,1.633909e+00,3.106163e-01,-0.381716e-01,1.43221,2.007994e+00,2.053545e+00,7.891387e-01,1.214766e-01,1.232024e+00,4.289248e-01,3.282060e+00,-1.959389e+00,5.446617e-01,-2.549571e-01,-2.564267e+00,-9.248252e-01,-3.081653e-01,-5.703988e-01,-1.293991e-01,4.973694e+00,-3.434379e-01,-1.380283e+00,4.498115e-01,-2.194023e-01,-7.609056e-02,6.926962e-01,3.025173e+00,1.225115e+00,-4.390540e-01,-1.967939e+00,6.067990e+00,-9.431182e-01,-1.080323e+00,4.858657e+00,-3.158694e+00,5.342106e-01,-6.174560e-01,4.278949e+00,1.434761e-01,5.881890e-01,4.307201e+00,2.995618e+00,-1.602819e+00,-4.700116e+00,-3.781894e+00,4.398818e-01,-2.486521e+00,3.825456e+00,-7.287003e+00,1.891789e-01,-3.076702e+00,3.401460e+00,3.117728e+00,1.271828e+00,-7.575880e-01,-3.955281e+00,-3.549420e+00,-1.950680e+00,4.185545e+00,6.633856e-02,-4.862587e+00,-3.171229e+00,5.363190e+00,8.934604e+00,3.284326e+00,4.810142e+00,3.277955e+00,5.086578e+00,1.855980e+00,-1.987295e+00,5.083860e+00,2.264808e+00,-5.762123e-01,3.295967e-02,-2.707252e+00,-5.881406e-01,-4.375442e+00,1.099283e+01,3.111975e-01,4.102009e+00,5.321727e-01,3.251275e+00,4.602224e+00,-1.201118e+00,-1.968421e+00,4.166239e-01,-2.088819e+00,1.546082e+00,2.200682e+00,4.852358e+00,3.016739e+00,9.841390e-01,-4.947384e-02,-3.364673e+00,-3.989378e+00,-6.419692e+00,5.191085e+00,-6.711466e+00,-4.232108e+00,9.578049e+00,-4.960075e+00,6.734879e-02,5.446707e+00,-1.360601e+00,-4.165223e-01,-5.524697e-01,2.883988e+00,6.089522e-01,-1.857194e+00,6.898488e-01,2.575902e+00,3.126794e+00,-6.801286e+00,-1.721459e+00,4.907785e+00,-4.157327e+00,3.712707e+00,-8.207942e+00,2.558770e+00,-3.621127e+00,4.847700e+00,-4.977816e+00,-1.348466e-01,8.086652e-01,5.607636e+00,-2.570728e+00,4.396520e+00,7.122537e+00,2.163478e-01,2.006025e+00,-6.276080e+00,1.666212e+00,5.101418e+00,2.710431e-01,2.308859e+00,-2.506426e+00,-2.600398e+00,-2.347812e+00,-6.222939e-01,9.489980e-01,-7.475396e-02,-2.991575e+00,-1.741864e+00,2.598526e+00,8.077144e-01,-3.187911e+00,6.412615e+00,1.991896e+00,-5.293677e+00,2.059473e+00,-7.621946e-01,2.516871e+00,2.879037e+00,-4.309212e+00,-4.592112e+00,-6.548691e+00,-3.375588e+00,-5.486407e-02,7.040420e-01,-3.713512e+00,-3.337896e-01,-2.710776e+00,-4.650604e+00,-7.243545e+00,-3.685247e+00,-1.083280e+01,-7.462294e-01,-5.487213e-01,-1.053556e+00,6.782904e-01,-3.063280e+00,5.669990e+00,-6.729556e+00,-5.939009e+00,-1.731180e+00,1.063611e+00,3.275798e+00,-1.211286e+01,9.489911e-01,-9.498630e+00,-1.127273e+00,-7.485883e+00,1.531899e+00,3.069147e-01,-3.725011e+00,1.735381e+00,5.634484e+00,-1.268195e+00,6.329459e+00,-9.959390e-02,-9.334362e-01,3.078773e-01,1.315052e+00,1.300378e+00,2.915159e+00,-8.123357e-02,-9.150178e-01,9.607442e-01,2.223099e+00,3.007240e+00,-1.495354e+00,1.901438e+00,2.856264e-01,-3.430402e+00,-2.359353e+00,-2.498281e+00,8.188639e+00,3.054233e+00,-1.162524e+00,2.288467e+00,9.600575e-01,-7.252222e+00,-3.887651e+00,-1.609879e+00,2.179292e+00,-5.585192e+00,3.250345e+00,8.524404e+00,-1.241254e+01,-4.376539e+00,1.767440e-02,4.589020e+00,7.213344e+00,-1.629210e+00,-7.649721e+00,1.329654e+00,-5.587428e+00,1.284069e+00,1.057744e+00,3.094442e-01,-5.058597e-01,4.902311e+00,7.567940e-01,-8.349228e+00,-1.378034e+00,-2.834166e-01,4.568659e+00,-7.612779e+00,-5.016349e-01,2.667397e+00,1.747587e-01,1.066581e+00,-2.525015e+00,4.017282e+00,2.934726e+00,-2.572435e+00,4.740898e+00,3.180701e+00,-5.006363e-01,3.211000e+00,6.973813e+00,5.261521e+00,-2.969950e+00,-4.730804e-01,1.490356e+00,8.656975e-01,2.404566e+00,-1.271306e+01,-1.678111e+00,-5.798974e-01,-5.239147e-02,3.994294e+00,-1.466720e-01,4.661478e+00,-3.669853e+00,2.824240e+00,4.752034e+00,-1.193278e+01,3.802017e-01,-8.816550e+00,-2.004147e+00,3.464639e+00,-7.932540e+00,-5.665270e+00,5.341964e+00,-1.856444e+00,3.767704e+00,2.094159e+00,-5.083992e+00,-9.680362e+00,2.117960e+00,1.810203e+00,1.871194e-01,-7.184265e+00,-4.159040e+00,-4.595828e+00,-8.480278e-01,-5.749642e+00,3.832951e+00,4.836513e+00,-4.278215e+00,3.510041e+00,-5.417238e+00,8.298303e+00,-7.970822e+00,-6.308041e+00,7.158563e-01,2.359358e+00,6.075790e+00,-6.826005e+00,-1.658422e+00,-5.372487e+00,3.625738e+00,3.606656e+00,2.667595e+00,-2.215293e+00,2.475873e+00,2.866556e+00,-3.305838e+00,2.512457e+00,8.482106e+00,6.487004e+00,1.688619e+00,3.729077e+00,-3.870335e+00,-3.987285e+00,-2.078456e+00,-6.474361e+00,-1.147600e+00,-4.274170e+00,-6.383891e-01,-1.700862e+00,2.922130e+00,-2.041512e+00,8.178325e-01,-6.882375e+00,-1.142741e+00,2.257451e+00,1.763709e+00,-3.812452e+00,-1.306668e+00,6.705246e+00,-5.158904e+00,6.119947e+00,4.126785e+00,-5.188898e-01,-1.977939e+00,3.344185e+00,-1.640737e+00,-6.993246e-01,2.182592e+00,-2.854662e+00,9.558266e-02,-3.236054e+00,5.609055e+00,-4.769000e+00,4.675493e+00,1.193283e-01,1.383967e+01,-5.880477e+00,-4.875417e-01,-6.278530e+00,-2.582043e-01,4.683166e-01,-2.972767e+00,1.358997e+01,3.167390e+00,3.929613e-01,2.852290e+00,5.885309e+00,2.574903e-01,-3.700218e+00,4.965009e-01,-3.238018e+00,5.940344e+00,-3.576156e+00,8.960790e+00,8.134748e+00,3.719045e+00,-1.665676e+00,-2.044653e+00,6.034864e+00,-6.245169e+00,4.741851e-02,4.183515e+00,1.379852e+00,3.450979e+00,-1.870875e+00,-1.035517e+01,-2.577533e+00,-9.765830e-02,-1.886556e+00,1.268050e+00,2.541175e+00,8.135890e+00,-5.702124e-02,5.692679e+00,-1.043058e+01,4.101508e+00,-1.995104e+00,8.145439e-01,6.349198e+00,-2.402096e+00,3.544194e+00,4.462910e+00,-8.132853e+00,-3.501196e+00,-1.261060e+00,-3.156412e+00,2.400902e+00,9.837739e-01,-3.220589e+00,2.156805e+00,-2.874818e-02,3.487826e-01,6.999742e+00,-3.531703e+00,-4.371512e-01,-3.210408e+00,6.322297e+00,-2.479741e-01,4.343348e+00,-4.437142e+00,-1.038985e+01,9.951091e-01,7.871011e-01,9.436210e+00,4.254338e+00,-1.138277e+00,-6.890702e+00,-2.230654e+00,-6.254442e+00,6.573852e-01,-9.758683e-01,-8.544015e+00,-3.837289e+00,5.501106e+00,5.515486e+00,-1.245664e+00,-6.201493e+00,8.608138e+00,4.326692e+00,-1.667451e+00,-6.265838e+00,7.579562e+00,-3.168849e+00,-2.829891e+00,1.333976e-01,5.752741e+00,6.717443e+00,9.168200e+00,1.597561e+00,7.658661e-01,2.094074e+00,-3.385709e+00,5.650523e+00,5.255718e+00,4.194054e+00,9.984224e+00,-4.993034e+00,-2.045212e+00,2.974039e+00,3.263424e-01,1.002583e+00,4.887874e+00,-2.297469e+00,5.264921e-02,5.828471e-01,4.935913e+00,4.567069e-01,-1.016908e+00,-3.195980e+00,-3.064826e+00,-3.623793e+00,-1.677771e-01,5.637015e+00,7.066110e+00,-1.388036e+00,2.488617e+00,-1.201026e-01,-2.799898e-01,-7.064959e+00,-5.723753e+00,1.079842e+00,3.259559e+00,1.494706e+00,-5.686538e-01,-2.865404e+00,-4.867026e+00,-1.369869e-01,1.209280e+00,1.558010e+00,-1.769743e+00,-1.995966e+00,-7.500511e-02,2.526998e+00,-5.361134e+00,1.378730e+00,4.239913e+00,-1.321548e+00,-3.028404e+00,-8.258905e+00,1.842130e+00,6.693688e+00};
+        double[] result = {0,-1.145232e-03,6.712909e-04,-3.650061e-04,8.144133e-03,-2.028834e-01,4.325592e+00,1.778990e+00,-9.738189e-01,2.914863e+00,4.625670e+00,3.774985e+00,6.412048e-01,-6.397910e-01,2.429660e+00,1.019273e+00,6.529764e+00,-4.328155e+00,1.587326e-01,-2.815395e-01,-4.844148e+00,-1.699221e+00,2.591101e-01,-5.985983e-01,-1.260872e-01,9.177906e+00,-7.674177e-01,-3.154319e+00,4.355714e-03,-1.311670e-01,1.086853e-01,1.240918e+00,4.663283e+00,1.525591e+00,-9.462387e-01,-3.244675e+00,8.930580e+00,-1.540933e+00,-1.477584e+00,6.423151e+00,-4.532109e+00,9.200771e-01,-1.070549e+00,6.182345e+00,1.870438e-01,7.091196e-01,5.664298e+00,4.008660e+00,-2.278300e+00,-6.430990e+00,-4.782385e+00,7.559431e-01,-2.935406e+00,5.163126e+00,-9.197744e+00,4.336167e-02,-3.666259e+00,4.284342e+00,3.952950e+00,1.439054e+00,-8.668607e-01,-4.771498e+00,-4.198757e+00,-2.255156e+00,4.866447e+00,-1.353871e-01,-5.848125e+00,-3.586798e+00,6.360068e+00,9.981677e+00,3.670659e+00,5.571982e+00,3.792485e+00,5.644697e+00,2.018416e+00,-2.440631e+00,6.078266e+00,2.592004e+00,-8.059353e-01,6.534365e-02,-3.070237e+00,-6.320791e-01,-4.929533e+00,1.236907e+01,2.710380e-01,4.334764e+00,8.538658e-01,3.385546e+00,4.843392e+00,-1.522997e+00,-2.223522e+00,5.909964e-01,-2.296611e+00,1.735444e+00,2.386663e+00,5.174195e+00,3.066875e+00,9.399236e-01,-7.489683e-02,-3.586128e+00,-4.137464e+00,-6.685145e+00,5.724723e+00,-7.304592e+00,-4.493758e+00,1.032715e+01,-5.826842e+00,-2.589634e-01,5.296977e+00,-1.554903e+00,-6.080229e-01,-5.207317e-01,3.073736e+00,6.062250e-01,-2.029511e+00,7.851389e-01,2.748394e+00,3.239532e+00,-7.226533e+00,-1.748024e+00,5.199150e+00,-4.446816e+00,3.836414e+00,-8.494255e+00,2.758450e+00,-3.645686e+00,4.983067e+00,-5.086956e+00,-2.241735e-01,8.342661e-01,5.818811e+00,-2.621320e+00,4.506850e+00,7.418929e+00,1.940240e-01,2.109284e+00,-6.457664e+00,1.744726e+00,5.119568e+00,2.158258e-01,2.329769e+00,-2.564036e+00,-2.608931e+00,-2.349517e+00,-6.181682e-01,9.467888e-01,-8.981105e-02,-3.029749e+00,-1.754641e+00,2.620752e+00,7.929375e-01,-3.216500e+00,6.501563e+00,1.997790e+00,-5.292272e+00,2.058339e+00,-6.911451e-01,2.555441e+00,2.921588e+00,-4.292377e+00,-4.701032e+00,-6.651084e+00,-3.425018e+00,-9.289794e-03,8.305904e-01,-3.749247e+00,-3.609430e-01,-2.716159e+00,-4.725702e+00,-7.339693e+00,-3.589032e+00,-1.080018e+01,-5.031004e-01,-3.676587e-01,-1.141766e+00,6.674021e-01,-3.019621e+00,5.580451e+00,-6.632674e+00,-5.854177e+00,-1.911994e+00,1.056410e+00,3.314172e+00,-1.207525e+01,1.050706e+00,-9.788493e+00,-1.605408e+00,-7.089798e+00,2.006714e+00,1.801533e-01,-3.844383e+00,1.625898e+00,5.749705e+00,-1.106993e+00,6.522239e+00,9.486398e-02,-9.607353e-01,3.417258e-01,1.328772e+00,1.318143e+00,2.949487e+00,-4.044447e-02,-9.278939e-01,9.944934e-01,2.251732e+00,3.024768e+00,-1.475238e+00,1.902920e+00,3.490702e-01,-3.462969e+00,-2.383715e+00,-2.508427e+00,8.165602e+00,3.129634e+00,-1.241960e+00,2.331049e+00,9.499026e-01,-7.239117e+00,-3.952925e+00,-1.491317e+00,2.251354e+00,-5.515696e+00,3.263587e+00,8.673782e+00,-1.237981e+01,-4.253246e+00,5.892398e-02,4.769566e+00,7.151571e+00,-1.694669e+00,-7.626578e+00,1.566971e+00,-5.659824e+00,1.262127e+00,9.278518e-01,4.140943e-01,-5.031820e-01,4.953299e+00,7.588297e-01,-8.330037e+00,-1.433181e+00,-4.676378e-01,4.753747e+00,-7.644284e+00,-5.987649e-01,2.657587e+00,3.481217e-01,1.067625e+00,-2.513933e+00,3.988478e+00,2.993604e+00,-2.540067e+00,4.702692e+00,3.242296e+00,-4.529561e-01,3.218345e+00,7.015370e+00,5.223388e+00,-3.030772e+00,-4.927134e-01,1.548966e+00,8.277164e-01,2.389188e+00,-1.261059e+01,-1.706569e+00,-5.618049e-01,-1.673632e-01,4.063490e+00,-5.650956e-02,4.688687e+00,-3.637568e+00,2.791329e+00,4.819106e+00,-1.199113e+01,1.610317e-01,-9.074054e+00,-1.920657e+00,3.551533e+00,-7.946825e+00,-5.567321e+00,5.525448e+00,-1.732386e+00,3.602129e+00,2.120435e+00,-4.912747e+00,-9.536617e+00,1.963228e+00,1.896941e+00,2.052092e-01,-7.253759e+00,-4.262683e+00,-4.718667e+00,-8.463998e-01,-5.740694e+00,3.851897e+00,4.860384e+00,-4.382090e+00,3.512529e+00,-5.570636e+00,8.453345e+00,-7.831959e+00,-6.078014e+00,7.676617e-01,2.508614e+00,6.060554e+00,-6.921532e+00,-1.784823e+00,-5.336305e+00,3.530468e+00,3.684109e+00,2.702318e+00,-2.252155e+00,2.564465e+00,2.875714e+00,-3.361014e+00,2.571348e+00,8.564147e+00,6.537697e+00,1.642932e+00,3.877917e+00,-3.972948e+00,-3.899810e+00,-2.130113e+00,-6.516929e+00,-1.135153e+00,-4.308537e+00,-5.446871e-01,-1.685854e+00,2.961681e+00,-2.033165e+00,8.597567e-01,-6.875942e+00,-1.111646e+00,2.184022e+00,1.820729e+00,-3.738894e+00,-1.295022e+00,6.635752e+00,-5.071652e+00,6.210003e+00,4.053425e+00,-5.921485e-01,-1.946577e+00,3.326141e+00,-1.596484e+00,-7.554612e-01,2.206324e+00,-2.852366e+00,5.660861e-02,-3.216249e+00,5.602577e+00,-4.742271e+00,4.695036e+00,1.170649e-01,1.396398e+01,-5.675921e+00,-6.189313e-01,-5.797253e+00,-2.714261e-04,4.114583e-01,-2.819026e+00,1.353096e+01,3.177567e+00,5.798262e-01,2.950829e+00,5.772202e+00,1.563067e-01,-3.863812e+00,4.421726e-01,-3.067972e+00,5.787052e+00,-3.501322e+00,8.859495e+00,8.152652e+00,3.490813e+00,-1.744665e+00,-2.197986e+00,5.929739e+00,-6.219281e+00,-1.908718e-02,4.086836e+00,1.423120e+00,3.572098e+00,-1.855752e+00,-1.041930e+01,-2.728980e+00,-2.127616e-01,-1.976158e+00,1.250741e+00,2.531691e+00,8.184168e+00,2.377821e-02,5.740594e+00,-1.045580e+01,4.224126e+00,-1.947085e+00,7.611733e-01,6.334082e+00,-2.319410e+00,3.537261e+00,4.554735e+00,-8.119148e+00,-3.512682e+00,-1.230155e+00,-3.200015e+00,2.417188e+00,1.000108e+00,-3.237766e+00,2.152887e+00,-2.807013e-02,3.395274e-01,6.982720e+00,-3.555731e+00,-4.762564e-01,-3.179971e+00,6.322914e+00,-2.601296e-01,4.362939e+00,-4.565139e+00,-1.050310e+01,8.472457e-01,8.449894e-01,9.299387e+00,4.188044e+00,-1.138349e+00,-6.776574e+00,-2.238330e+00,-6.319358e+00,7.832209e-01,-9.340360e-01,-8.582595e+00,-3.824387e+00,5.529635e+00,5.513343e+00,-1.099106e+00,-6.153996e+00,8.629339e+00,4.485832e+00,-1.619169e+00,-6.212311e+00,7.561383e+00,-3.379408e+00,-2.704568e+00,1.370970e-01,5.719368e+00,6.596437e+00,9.073606e+00,1.691819e+00,6.645324e-01,2.100249e+00,-3.430365e+00,5.748249e+00,5.259474e+00,4.296136e+00,1.007277e+01,-5.155287e+00,-1.979243e+00,3.098119e+00,2.987865e-01,1.028644e+00,4.737405e+00,-2.293267e+00,1.101671e-01,5.105804e-01,4.858547e+00,4.225842e-01,-9.796588e-01,-3.216708e+00,-2.991805e+00,-3.575280e+00,-1.297814e-01,5.571457e+00,7.049356e+00,-1.407684e+00,2.506835e+00,-2.178931e-01,-2.845864e-01,-6.929764e+00,-5.599349e+00,1.171110e+00,3.288733e+00,1.583220e+00,-5.605546e-01,-2.907751e+00,-4.901159e+00,-1.764410e-01,1.231616e+00,1.650142e+00,-1.752448e+00,-2.016070e+00,-1.095978e-01,2.540034e+00,-5.328199e+00,1.343955e+00,4.155428e+00,-1.261429e+00,-3.036951e+00,-8.301551e+00,1.861866e+00,6.610725e+00};
 
-        NLMSFilter filt1 = new NLMSFilter(mu, length, NLMSFilter.WeightsFillMethod.ZEROS);
+        LMS filt1 = new LMS(mu, length, LMS.WeightsFillMethod.ZEROS);
         filt1.filter(this.desired1, this.signal1);
 
-        NLMSFilter filt2 = new NLMSFilter(mu, weights);
+        LMS filt2 = new LMS(mu, weights);
         filt2.filter(this.desired1, this.signal1);
 
-        Assertions.assertArrayEquals(result, filt1.getOutput(), 0.1);
-        Assertions.assertArrayEquals(result, filt2.getOutput(), 0.1);
+        Assertions.assertArrayEquals(result, filt1.getOutput(), 0.05);
+        Assertions.assertArrayEquals(result, filt2.getOutput(), 0.05);
 
         // Save plot result
         Plotting fig = new Plotting(600, 500, "Sample Figure", "Time", "Signal");
@@ -61,23 +61,23 @@ public class TestNLMSFilter {
     }
 
     @Test
-    public void NLMSTest2() throws IOException {
-        double mu = 1;
+    public void LMSTest2() throws IOException {
+        double mu = 0.1;
         int length = 20;
-        String outputFileName = "test_outputs/NLMSTest2.png";
+        String outputFileName = "test_outputs/LMSTest2.png";
         double[] weights = new double[length];
         Arrays.fill(weights, 0);
 
         // Results calculated with MATLAB R2020b Update 4 9.9.0.1570001
-        double[] result = {0,0,-1.464684e-02,-3.050340e-02,1.009006e-01,2.026721e-01,3.003929e-01,3.969671e-01,6.112817e-01,5.681300e-01,4.699671e-01,5.033567e-01,4.809130e-01,5.933191e-01,6.075221e-01,6.351595e-01,6.905352e-01,8.033503e-01,8.276656e-01,8.673143e-01,8.570127e-01,8.023211e-01,8.633987e-01,9.248569e-01,9.637966e-01,9.827972e-01,9.624284e-01,9.467327e-01,9.735977e-01,9.524028e-01,1.009172e+00,9.716751e-01,9.917334e-01,9.589005e-01,9.347995e-01,1.035212e+00,1.012071e+00,1.017354e+00,9.849985e-01,8.865540e-01,8.990965e-01,8.588833e-01,8.794055e-01,8.440648e-01,8.099807e-01,7.579768e-01,7.292700e-01,7.318557e-01,7.005843e-01,6.729715e-01,5.680565e-01,5.290374e-01,4.358606e-01,4.269830e-01,4.026828e-01,4.253858e-01,3.404978e-01,3.118064e-01,2.004642e-01,2.010205e-01,1.019231e-01,8.076927e-02,8.943492e-03,1.912048e-02,-2.131381e-02,-1.053808e-01,-2.193688e-01,-2.805912e-01,-3.246482e-01,-2.661311e-01,-3.374654e-01,-3.469923e-01,-4.708570e-01,-4.789979e-01,-5.684074e-01,-6.147989e-01,-6.902257e-01,-6.592337e-01,-6.667857e-01,-6.875661e-01,-7.049422e-01,-7.757338e-01,-7.970520e-01,-8.367363e-01,-9.095045e-01,-9.132971e-01,-9.288962e-01,-9.645081e-01,-9.593527e-01,-1.004585e+00,-1.024342e+00,-9.933125e-01,-9.585740e-01,-9.000228e-01,-9.756844e-01,-9.965896e-01,-1.046049e+00,-1.052095e+00,-1.011315e+00,-1.021284e+00,-9.099159e-01,-9.220003e-01,-8.560339e-01,-8.854484e-01,-8.333195e-01,-8.473113e-01,-7.845335e-01,-7.701786e-01,-6.917745e-01,-7.072162e-01,-7.405269e-01,-7.058027e-01,-6.025181e-01,-5.416441e-01,-4.696715e-01,-4.609140e-01,-3.848706e-01,-4.157596e-01,-3.555704e-01,-3.639391e-01,-1.693411e-01,-1.398827e-01,-6.372772e-02,-1.149869e-01,-8.269806e-02,-3.428019e-02,6.618312e-02,1.214643e-01,1.824173e-01,1.770671e-01,2.168433e-01,2.811375e-01,3.174772e-01,4.194565e-01,4.654725e-01,5.382752e-01,5.323777e-01,6.017736e-01,6.160955e-01,6.658676e-01,6.799298e-01,7.014578e-01,7.186356e-01,7.794764e-01,7.997706e-01,8.440737e-01,8.744085e-01,8.976070e-01,9.014594e-01,9.792899e-01,9.963967e-01,1.013336e+00,9.905045e-01,9.636824e-01,1.003396e+00,1.002226e+00,9.883768e-01,1.013236e+00,9.986621e-01,9.839814e-01,9.669910e-01,9.362656e-01,9.182184e-01,9.949046e-01,9.812303e-01,9.561725e-01,8.500140e-01,7.847664e-01,7.643933e-01,8.069000e-01,7.497820e-01,6.827738e-01,6.461022e-01,6.590590e-01,6.768550e-01,6.446849e-01,5.577612e-01,5.035992e-01,4.377505e-01,4.145172e-01,3.300165e-01,3.122146e-01,2.255852e-01,2.135534e-01,1.830232e-01,1.611905e-01,6.475812e-02,3.625002e-02,-3.478883e-02,-7.640946e-02,-1.404756e-01,-2.197841e-01,-2.634662e-01,-2.900750e-01,-3.273800e-01,-3.594520e-01,-3.741371e-01,-4.388935e-01,-5.188425e-01,-5.674414e-01,-6.260209e-01,-6.664343e-01,-6.756713e-01,-7.133605e-01,-7.202247e-01,-7.507683e-01,-8.093898e-01,-8.576479e-01,-8.774264e-01,-9.082206e-01,-9.196294e-01,-8.854464e-01,-9.416393e-01,-9.539930e-01,-1.005153e+00,-1.013648e+00,-9.969378e-01,-9.872919e-01,-9.504808e-01,-9.648023e-01,-9.873498e-01,-9.810762e-01,-9.516815e-01,-9.585073e-01,-1.000226e+00,-1.002107e+00,-9.611049e-01,-8.904119e-01,-8.707828e-01,-8.401696e-01,-8.241723e-01,-8.117148e-01,-8.092040e-01,-7.876918e-01,-7.285293e-01,-6.753520e-01,-6.168369e-01,-6.412941e-01,-5.584384e-01,-4.928999e-01,-4.515198e-01,-3.680923e-01,-3.544106e-01,-3.145244e-01,-2.778006e-01,-2.678590e-01,-2.176933e-01,-1.165724e-01,-5.908011e-02,-5.742621e-03,5.015560e-03,4.663056e-02,1.105865e-01,1.770275e-01,2.109730e-01,2.581605e-01,3.518022e-01,3.383249e-01,4.454946e-01,4.608120e-01,5.331870e-01,5.120001e-01,5.428226e-01,5.731997e-01,6.752587e-01,6.952617e-01,7.420430e-01,7.607246e-01,8.100363e-01,8.686892e-01,8.779656e-01,9.206834e-01,9.276194e-01,9.197045e-01,9.225568e-01,9.115531e-01,9.262769e-01,9.708645e-01,1.005114e+00,1.025677e+00,1.036049e+00,1.024162e+00,1.015339e+00,1.011064e+00,1.000918e+00,9.729713e-01,9.438265e-01,9.105225e-01,9.261600e-01,9.011180e-01,8.982200e-01,8.343253e-01,8.577838e-01,8.588137e-01,8.390769e-01,7.801659e-01,6.779533e-01,6.663263e-01,6.109206e-01,5.780020e-01,5.558283e-01,5.419340e-01,5.362510e-01,5.125317e-01,4.573956e-01,3.719366e-01,3.149544e-01,2.392571e-01,2.233174e-01,1.704077e-01,1.382144e-01,6.084495e-02,5.630310e-03,-3.192565e-02,-1.262579e-01,-1.687702e-01,-2.419321e-01,-2.561134e-01,-2.646183e-01,-2.843457e-01,-3.557995e-01,-4.298592e-01,-4.822928e-01,-5.247039e-01,-5.454082e-01,-5.894432e-01,-6.418258e-01,-6.844834e-01,-7.472443e-01,-7.489838e-01,-7.727868e-01,-8.039127e-01,-8.423198e-01,-8.807743e-01,-8.892333e-01,-9.012026e-01,-8.799854e-01,-8.901755e-01,-9.614449e-01,-9.991126e-01,-1.042938e+00,-1.020218e+00,-9.917929e-01,-9.726552e-01,-9.774707e-01,-9.455359e-01,-9.808791e-01,-9.809215e-01,-9.849019e-01,-9.704188e-01,-9.405357e-01,-9.290523e-01,-9.199446e-01,-9.394477e-01,-8.668881e-01,-8.393882e-01,-7.823969e-01,-7.579908e-01,-7.402406e-01,-7.159704e-01,-6.914335e-01,-6.567817e-01,-5.811894e-01,-5.340177e-01,-4.914639e-01,-4.549523e-01,-4.222135e-01,-3.925954e-01,-3.572671e-01,-3.023375e-01,-2.509584e-01,-1.817777e-01,-1.409803e-01,-8.041902e-02,-9.527726e-03,2.326260e-02,6.176225e-02,1.204125e-01,1.472628e-01,2.164228e-01,3.094312e-01,3.755373e-01,4.017731e-01,4.024140e-01,4.084592e-01,4.867729e-01,5.569053e-01,6.088867e-01,6.169676e-01,6.478679e-01,6.731206e-01,7.130883e-01,7.494650e-01,7.753021e-01,8.149064e-01,8.592381e-01,8.785931e-01,9.221165e-01,9.493284e-01,9.831738e-01,9.964210e-01,9.958571e-01,9.664767e-01,9.541671e-01,9.585899e-01,9.927935e-01,1.037092e+00,1.036374e+00,1.037535e+00,1.012233e+00,1.004912e+00,9.669879e-01,9.262354e-01,8.789210e-01,8.753901e-01,8.962476e-01,8.904618e-01,8.642427e-01,8.156298e-01,7.750435e-01,7.401326e-01,7.170723e-01,6.886882e-01,6.654450e-01,6.394785e-01,5.940308e-01,5.682033e-01,5.115298e-01,4.756436e-01,3.930142e-01,3.375380e-01,2.742914e-01,2.378928e-01,2.048860e-01,1.320164e-01,6.579426e-02,-7.263374e-04,-1.041441e-02,-1.196462e-02,-4.641843e-02,-1.042689e-01,-1.656947e-01,-2.460830e-01,-3.050740e-01,-3.682965e-01,-3.888843e-01,-4.094408e-01,-4.319043e-01,-4.860956e-01,-5.526914e-01,-6.103449e-01,-6.924884e-01,-7.353731e-01,-7.375963e-01,-7.438390e-01,-7.670775e-01,-8.020245e-01,-8.316613e-01,-8.487295e-01,-8.663955e-01,-9.006627e-01,-8.972899e-01,-9.454631e-01,-9.510715e-01,-1.011151e+00,-1.009738e+00,-1.015778e+00,-9.825025e-01,-9.618150e-01,-9.523096e-01,-9.644166e-01,-9.828073e-01,-9.683261e-01,-9.665646e-01,-9.536391e-01,-9.545100e-01,-9.487398e-01,-9.086774e-01,-9.007276e-01,-8.660309e-01,-8.403631e-01,-7.956307e-01,-7.710204e-01,-7.202964e-01,-6.784244e-01,-6.494246e-01,-6.182833e-01,-5.968949e-01,-5.525933e-01,-5.050590e-01,-4.671168e-01,-4.145648e-01,-3.578243e-01,-3.043296e-01,-2.342678e-01,-2.069498e-01,-1.890837e-01,-1.711874e-01,-1.054890e-01,-2.068774e-02,5.349355e-02,1.119168e-01,1.507163e-01,1.973047e-01,2.380954e-01,2.790983e-01,3.172046e-01,4.100770e-01,4.380879e-01,4.505282e-01,4.564999e-01,4.949939e-01,5.744880e-01,6.447323e-01,7.076177e-01,7.520261e-01,7.943515e-01,8.161258e-01,8.437794e-01,8.582108e-01,8.521532e-01,8.271387e-01,8.490405e-01,8.848551e-01,9.387228e-01,9.736833e-01,1.021284e+00,1.044160e+00,1.027504e+00,9.924164e-01,9.755787e-01,9.824471e-01,9.801136e-01,9.679819e-01,9.695968e-01,9.748644e-01,9.597449e-01,9.627868e-01,9.520147e-01,9.393971e-01,9.164480e-01,8.776208e-01,8.629182e-01,8.200112e-01,7.915453e-01,7.591089e-01,7.279873e-01,6.739925e-01,6.140943e-01,5.740501e-01,5.483552e-01,5.194273e-01,4.951696e-01,4.583928e-01,4.235507e-01,3.636172e-01,3.048594e-01,2.454295e-01,1.892501e-01,1.462516e-01,1.037386e-01,4.768701e-02,1.419485e-02,-2.264970e-02,-7.460718e-02,-1.356748e-01,-2.132383e-01,-2.586728e-01,-2.890596e-01,-3.421419e-01,-3.928626e-01,-4.268616e-01,-4.487334e-01,-4.854054e-01,-5.364625e-01,-6.285598e-01,-6.997013e-01,-7.449014e-01,-7.370123e-01,-7.207425e-01,-7.358258e-01,-7.800569e-01,-8.099525e-01,-8.717551e-01,-9.193518e-01,-9.472980e-01,-9.426050e-01,-9.266012e-01,-9.291032e-01,-9.617204e-01,-9.699807e-01,-9.792737e-01,-9.854583e-01,-9.706272e-01,-9.701598e-01,-9.643458e-01,-9.736478e-01,-9.609093e-01,-9.524552e-01,-9.386683e-01,-9.272081e-01,-9.067775e-01,-8.928144e-01,-8.929571e-01,-8.535957e-01,-8.407787e-01,-8.063322e-01,-7.728725e-01,-7.426145e-01,-6.799167e-01,-6.375903e-01,-6.082949e-01,-5.899550e-01,-5.587702e-01,-5.371217e-01,-4.769524e-01,-4.155304e-01,-3.456193e-01,-2.810931e-01,-2.082495e-01,-1.634635e-01,-1.326916e-01,-1.075958e-01,-7.866530e-02,-4.585720e-02,-6.490557e-03,7.078067e-02,1.408038e-01,2.136540e-01,2.531899e-01,2.914978e-01,3.397562e-01,3.717994e-01,4.271035e-01,4.618868e-01,5.126427e-01,5.668339e-01,5.885838e-01,6.089128e-01,6.493585e-01,6.983148e-01,7.520529e-01,7.905303e-01,8.440931e-01,8.600871e-01,8.972328e-01,9.346641e-01,9.498420e-01,9.597942e-01,9.580876e-01,9.484127e-01,9.547562e-01,9.794860e-01,1.023729e+00,1.033970e+00,1.003798e+00,9.844318e-01,9.738874e-01,9.808847e-01,9.712487e-01,9.747666e-01,9.541543e-01,9.403604e-01,9.104584e-01,8.844124e-01,8.798929e-01,8.819068e-01,8.600346e-01,8.351292e-01,8.126123e-01,7.743614e-01,7.288333e-01,6.831540e-01,6.505626e-01,6.247192e-01,5.937462e-01,5.548032e-01,4.823376e-01,4.023043e-01,3.270822e-01,2.855564e-01,2.386232e-01,2.199935e-01,1.909841e-01,1.458550e-01,9.404637e-02,3.763544e-02,-1.050759e-02,-6.240963e-02,-1.121085e-01,-1.578637e-01,-1.913471e-01,-2.302605e-01,-2.899280e-01,-3.595164e-01,-4.330714e-01,-4.763484e-01,-5.059615e-01,-5.243512e-01,-5.623044e-01,-5.878939e-01,-6.273465e-01,-6.591757e-01,-7.234108e-01,-7.814543e-01,-8.176464e-01,-8.365592e-01,-8.622430e-01,-9.099925e-01,-9.376486e-01,-9.603841e-01,-9.512281e-01,-9.503958e-01,-9.486049e-01,-9.693385e-01,-9.767651e-01,-9.704279e-01,-9.753758e-01,-9.745483e-01,-9.938238e-01,-9.921366e-01,-9.902051e-01,-9.721736e-01,-9.729530e-01,-9.612822e-01,-9.489724e-01,-9.347267e-01,-9.255540e-01,-8.930777e-01,-8.562757e-01,-8.226057e-01,-7.945951e-01,-7.737230e-01,-7.596134e-01,-7.366309e-01,-6.947201e-01,-6.388726e-01,-5.733432e-01,-5.065951e-01,-4.454368e-01,-3.930759e-01,-3.629352e-01,-3.606702e-01,-3.413609e-01,-3.015082e-01,-2.406339e-01,-1.659311e-01,-8.873567e-02,-2.482876e-02,-5.785654e-03,3.629458e-02,9.694077e-02,1.686322e-01,1.965290e-01,2.322972e-01,2.719655e-01,3.188807e-01,3.704050e-01,4.059889e-01,4.413928e-01,4.810957e-01,5.485726e-01,6.073708e-01,6.679085e-01,7.039701e-01,7.393929e-01,7.612042e-01,7.769206e-01,7.871081e-01,8.072691e-01,8.394428e-01,8.686343e-01,8.942261e-01,9.302967e-01,9.530080e-01,9.449191e-01,9.554789e-01,9.786878e-01,1.004308e+00,1.022318e+00,1.042454e+00,1.023561e+00,9.899560e-01,9.553720e-01,9.492825e-01,9.783279e-01,9.968760e-01,9.878316e-01,9.490173e-01,8.984290e-01,8.469066e-01,8.245737e-01,7.983193e-01,7.929232e-01,7.921500e-01,7.770040e-01,7.521871e-01,6.852679e-01,6.308126e-01,5.893535e-01,5.641972e-01,5.104429e-01,4.468738e-01,3.873715e-01,3.611934e-01,3.350482e-01,2.927391e-01,2.382651e-01,1.838379e-01,1.191260e-01,6.883563e-02,1.530147e-02,-3.759718e-02,-8.359813e-02,-1.231456e-01,-1.593013e-01,-2.030254e-01,-2.656259e-01,-3.168791e-01,-3.674284e-01,-4.186915e-01,-4.710618e-01,-5.097695e-01,-5.567438e-01,-6.018457e-01,-6.383040e-01,-6.685642e-01,-6.971123e-01,-7.200785e-01,-7.506225e-01,-7.795891e-01,-8.049571e-01,-8.440955e-01,-8.797558e-01,-9.084814e-01,-9.309558e-01,-9.588474e-01,-9.744598e-01,-9.863411e-01,-1.005089e+00,-1.012430e+00,-1.003639e+00,-9.883554e-01,-9.858724e-01,-1.000122e+00,-1.026833e+00,-1.033970e+00,-1.007991e+00,-9.679244e-01,-9.370735e-01,-9.155982e-01,-9.115947e-01,-8.982863e-01,-8.598183e-01,-8.388420e-01,-8.369608e-01,-8.303016e-01,-7.915160e-01,-7.270287e-01,-6.628486e-01,-6.279236e-01,-6.083094e-01,-5.787608e-01,-5.531381e-01,-5.202991e-01,-4.760659e-01,-4.204838e-01,-3.614506e-01,-3.041730e-01,-2.499482e-01,-2.017696e-01,-1.556450e-01,-1.039483e-01,-4.528145e-02,1.114066e-02,4.671488e-02,6.799269e-02,9.183305e-02,1.467599e-01,2.390449e-01,3.178145e-01,3.646566e-01,3.808189e-01,4.014032e-01,4.481198e-01,4.919899e-01,5.296894e-01,5.668230e-01,6.075700e-01,6.622092e-01,7.031172e-01,7.711380e-01,8.138998e-01,8.413223e-01,8.387783e-01,8.307712e-01,8.371635e-01,8.671510e-01,9.422215e-01,9.661516e-01,9.925120e-01,9.827677e-01,9.833883e-01,9.979668e-01,1.005936e+00,9.762898e-01,9.702779e-01,9.699125e-01,9.580696e-01,9.706971e-01,9.636590e-01,9.846213e-01,9.605170e-01,9.480837e-01,8.844597e-01,8.444904e-01,8.026312e-01,7.968497e-01,8.005349e-01,8.008438e-01,8.079036e-01,7.656939e-01,7.049656e-01,6.452302e-01,5.789841e-01,5.286895e-01,4.943558e-01,4.431637e-01,4.040874e-01,3.515772e-01,3.048863e-01,2.556779e-01,2.122872e-01,1.647987e-01,1.027013e-01,6.015542e-02,2.676285e-02,-1.375710e-02,-5.235819e-02,-1.179040e-01,-1.648261e-01,-2.278856e-01,-2.717471e-01,-3.331742e-01,-3.809621e-01,-4.233155e-01,-4.738320e-01,-5.284469e-01,-5.723137e-01,-6.084926e-01,-6.400212e-01,-6.704634e-01,-7.093037e-01,-7.503688e-01,-7.769461e-01,-7.867170e-01,-8.057686e-01,-8.334330e-01,-8.532240e-01,-8.833103e-01,-9.200246e-01,-9.719369e-01,-9.986564e-01,-1.010844e+00,-1.008304e+00,-1.004936e+00,-9.969730e-01,-9.811610e-01,-9.775412e-01,-9.890582e-01,-1.005461e+00,-1.008121e+00,-1.002415e+00,-9.777434e-01,-9.477130e-01,-9.154007e-01,-9.039231e-01,-8.860051e-01,-8.558321e-01,-8.254632e-01,-8.123992e-01,-8.018300e-01,-7.709917e-01,-7.352341e-01,-6.880419e-01,-6.391829e-01,-5.965123e-01,-5.660961e-01,-5.483255e-01,-5.337457e-01,-5.011606e-01,-4.400708e-01,-3.744731e-01,-2.937632e-01,-2.171839e-01,-1.514674e-01,-1.014354e-01,-7.087117e-02,-4.418094e-02,-1.841503e-02};
+        double[] result = {0,0,-8.421734e-05,-2.225732e-04,6.589808e-04,2.152792e-03,5.844287e-03,1.461282e-02,4.124764e-02,8.911210e-02,1.264123e-01,2.138494e-01,2.954630e-01,3.746900e-01,4.693800e-01,5.471038e-01,6.224386e-01,7.202252e-01,7.921595e-01,8.399704e-01,8.577754e-01,8.450263e-01,8.993684e-01,9.437290e-01,9.391693e-01,9.644819e-01,9.706059e-01,9.651871e-01,9.933987e-01,9.756522e-01,1.023895e+00,9.646808e-01,1.006814e+00,9.827173e-01,9.568060e-01,1.048735e+00,9.403563e-01,9.845438e-01,9.597129e-01,8.948503e-01,9.380087e-01,8.648943e-01,8.890165e-01,8.215463e-01,8.041923e-01,7.754266e-01,7.423613e-01,7.166920e-01,6.852841e-01,6.416320e-01,5.698700e-01,5.578308e-01,4.816138e-01,4.503566e-01,4.205737e-01,4.027192e-01,3.127503e-01,2.854733e-01,2.332498e-01,2.007975e-01,1.229624e-01,8.419749e-02,5.036764e-02,1.656973e-02,-1.964423e-02,-8.100072e-02,-1.620753e-01,-2.120238e-01,-2.612091e-01,-2.519046e-01,-3.052466e-01,-3.327289e-01,-3.785539e-01,-4.092054e-01,-4.654936e-01,-5.227459e-01,-5.907126e-01,-6.070044e-01,-6.475931e-01,-6.821977e-01,-6.972190e-01,-7.412858e-01,-7.840075e-01,-7.894104e-01,-8.622249e-01,-8.854600e-01,-9.125950e-01,-9.601434e-01,-9.696972e-01,-9.864560e-01,-1.013955e+00,-9.733674e-01,-9.829667e-01,-9.373614e-01,-1.045728e+00,-9.857252e-01,-1.012265e+00,-1.018864e+00,-9.672050e-01,-1.006103e+00,-8.837979e-01,-9.942936e-01,-8.501631e-01,-9.364704e-01,-8.141755e-01,-8.869306e-01,-7.445110e-01,-8.176325e-01,-6.745537e-01,-7.792587e-01,-6.917730e-01,-6.499351e-01,-5.598224e-01,-5.865262e-01,-4.875195e-01,-4.757663e-01,-3.977594e-01,-4.408190e-01,-3.395045e-01,-3.196043e-01,-1.752398e-01,-1.589024e-01,-1.120863e-01,-1.001688e-01,-6.290082e-02,-2.694137e-03,7.158764e-02,1.206461e-01,1.869603e-01,1.941906e-01,2.209839e-01,2.691190e-01,2.955287e-01,3.688258e-01,4.411898e-01,5.007723e-01,5.359775e-01,6.033917e-01,6.599248e-01,6.809604e-01,7.367389e-01,7.649146e-01,7.643270e-01,7.859514e-01,8.085323e-01,8.236985e-01,8.585105e-01,8.897107e-01,8.973004e-01,9.533202e-01,9.916241e-01,9.773469e-01,9.843565e-01,9.825344e-01,1.029810e+00,9.973045e-01,9.798541e-01,1.046580e+00,9.784488e-01,9.816828e-01,9.841052e-01,9.498581e-01,9.364896e-01,1.006884e+00,9.074319e-01,9.103322e-01,8.382301e-01,8.592512e-01,8.009884e-01,8.330714e-01,7.114380e-01,7.185313e-01,6.780895e-01,6.743494e-01,6.363194e-01,5.867828e-01,5.385896e-01,5.116244e-01,4.444757e-01,4.227769e-01,3.380537e-01,3.136242e-01,2.470887e-01,2.149605e-01,1.941356e-01,1.505306e-01,5.847753e-02,2.642560e-02,-7.937750e-03,-7.050859e-02,-1.319663e-01,-2.067650e-01,-2.648365e-01,-3.092713e-01,-3.470872e-01,-3.765063e-01,-3.911005e-01,-4.342583e-01,-4.952116e-01,-5.301461e-01,-5.859761e-01,-6.547368e-01,-6.840993e-01,-7.181332e-01,-7.526005e-01,-7.690735e-01,-8.124359e-01,-8.560350e-01,-8.788863e-01,-8.971980e-01,-9.270571e-01,-9.004693e-01,-9.563320e-01,-9.757491e-01,-9.844010e-01,-1.001204e+00,-9.765484e-01,-9.973347e-01,-9.643397e-01,-1.005196e+00,-1.009935e+00,-9.556532e-01,-9.689751e-01,-9.886576e-01,-9.816417e-01,-9.401485e-01,-9.349041e-01,-9.056250e-01,-8.937791e-01,-8.509020e-01,-8.409861e-01,-8.187100e-01,-7.951203e-01,-7.472107e-01,-7.206901e-01,-6.767238e-01,-6.317758e-01,-6.549291e-01,-5.198863e-01,-4.965653e-01,-4.946217e-01,-3.820307e-01,-3.612807e-01,-3.276564e-01,-2.813178e-01,-2.601282e-01,-1.895275e-01,-9.605430e-02,-6.880423e-02,-2.467027e-02,1.364135e-02,5.878847e-02,1.170671e-01,1.728099e-01,2.141850e-01,2.580805e-01,3.535097e-01,3.417784e-01,4.206291e-01,4.927149e-01,5.358389e-01,5.514429e-01,5.694694e-01,6.124139e-01,6.576334e-01,6.666003e-01,7.102405e-01,7.528026e-01,7.867638e-01,8.528986e-01,8.723686e-01,9.171589e-01,9.351992e-01,9.265856e-01,9.345255e-01,9.408374e-01,9.488541e-01,9.868976e-01,1.002787e+00,9.934586e-01,1.023298e+00,9.979394e-01,1.001781e+00,1.011742e+00,9.963056e-01,9.639632e-01,9.654237e-01,9.305897e-01,9.719637e-01,8.618995e-01,9.378034e-01,8.131538e-01,9.146293e-01,7.915052e-01,8.202302e-01,7.384320e-01,6.949305e-01,7.251464e-01,6.054936e-01,5.971725e-01,5.948897e-01,5.247877e-01,5.079391e-01,4.776991e-01,4.281668e-01,3.584797e-01,3.163586e-01,2.724832e-01,2.382452e-01,1.801833e-01,1.389848e-01,6.632108e-02,1.426231e-02,-1.263156e-02,-1.031834e-01,-1.661229e-01,-2.369502e-01,-2.768469e-01,-3.082853e-01,-3.412627e-01,-3.918780e-01,-4.487169e-01,-4.764249e-01,-5.143584e-01,-5.412090e-01,-5.746896e-01,-6.240749e-01,-6.689980e-01,-7.297382e-01,-7.419127e-01,-7.622582e-01,-8.256610e-01,-8.402735e-01,-8.746796e-01,-8.969166e-01,-9.096716e-01,-8.931263e-01,-9.094198e-01,-9.974525e-01,-9.633614e-01,-1.002445e+00,-1.001476e+00,-9.793088e-01,-9.934907e-01,-9.957366e-01,-9.695444e-01,-1.008425e+00,-9.751142e-01,-9.754451e-01,-9.548743e-01,-9.393016e-01,-9.352030e-01,-9.197744e-01,-9.240436e-01,-8.105521e-01,-8.817289e-01,-7.687051e-01,-7.860396e-01,-7.457859e-01,-7.188717e-01,-6.882551e-01,-6.268272e-01,-5.775043e-01,-5.611212e-01,-5.051343e-01,-4.445606e-01,-4.384603e-01,-3.874586e-01,-3.370123e-01,-2.800054e-01,-2.521779e-01,-1.838695e-01,-1.440583e-01,-1.022914e-01,-1.876293e-02,9.108004e-03,5.667121e-02,1.239948e-01,1.489845e-01,2.059194e-01,3.068867e-01,3.666373e-01,4.053829e-01,4.372038e-01,4.640788e-01,5.201840e-01,5.770562e-01,6.044858e-01,6.194061e-01,6.445410e-01,6.675668e-01,6.893585e-01,7.199744e-01,7.456168e-01,7.846558e-01,8.263418e-01,8.481880e-01,9.019312e-01,9.301639e-01,9.650652e-01,9.934024e-01,9.905705e-01,9.659061e-01,9.726788e-01,9.932355e-01,1.015825e+00,1.028289e+00,9.986529e-01,1.013783e+00,9.985134e-01,9.871028e-01,9.671188e-01,9.323164e-01,9.188887e-01,9.179222e-01,9.102446e-01,8.446514e-01,8.625797e-01,7.972256e-01,7.935411e-01,7.420765e-01,7.371951e-01,6.737123e-01,6.575762e-01,6.183164e-01,5.779419e-01,5.611681e-01,4.882108e-01,4.718018e-01,3.907542e-01,3.528987e-01,3.049163e-01,2.474227e-01,2.054257e-01,1.170881e-01,6.561660e-02,1.270941e-02,-1.799934e-02,-3.156524e-02,-9.211088e-02,-1.356294e-01,-1.723406e-01,-2.435035e-01,-2.938968e-01,-3.458164e-01,-3.757705e-01,-4.027477e-01,-4.310600e-01,-4.648734e-01,-5.223480e-01,-5.643465e-01,-6.572861e-01,-7.170610e-01,-7.326716e-01,-7.711245e-01,-8.257618e-01,-8.558975e-01,-8.497954e-01,-8.682151e-01,-8.789295e-01,-8.980724e-01,-9.007347e-01,-9.514243e-01,-9.674714e-01,-9.936011e-01,-9.946542e-01,-9.909346e-01,-9.821229e-01,-9.825998e-01,-9.909890e-01,-9.982809e-01,-9.884749e-01,-9.686482e-01,-9.753974e-01,-9.477613e-01,-9.512724e-01,-9.406196e-01,-8.895945e-01,-9.087836e-01,-8.531581e-01,-8.393844e-01,-7.897321e-01,-7.916637e-01,-7.126132e-01,-7.099984e-01,-6.670369e-01,-6.322339e-01,-5.859510e-01,-5.441524e-01,-4.977146e-01,-4.755582e-01,-4.085518e-01,-3.551249e-01,-3.197131e-01,-2.451691e-01,-2.214741e-01,-2.016915e-01,-1.571997e-01,-8.912463e-02,-1.419312e-02,2.669910e-02,8.743241e-02,1.382909e-01,1.895374e-01,2.413824e-01,2.888357e-01,3.351898e-01,4.165011e-01,4.517096e-01,4.463799e-01,4.647241e-01,4.899585e-01,5.307483e-01,5.792149e-01,6.439553e-01,7.006852e-01,7.529726e-01,7.922434e-01,8.586771e-01,8.925712e-01,8.670983e-01,8.467267e-01,8.985631e-01,9.066144e-01,9.267223e-01,9.702619e-01,1.008029e+00,1.013212e+00,9.658125e-01,1.000049e+00,9.923219e-01,9.984692e-01,9.859692e-01,9.782948e-01,9.924890e-01,9.630405e-01,9.525492e-01,9.838014e-01,9.258495e-01,9.285716e-01,8.962234e-01,8.591882e-01,8.723143e-01,7.867010e-01,8.183410e-01,7.436989e-01,7.372488e-01,6.574232e-01,6.360779e-01,6.019888e-01,5.619518e-01,5.176193e-01,4.901491e-01,4.315128e-01,4.066963e-01,3.350849e-01,2.967212e-01,2.564823e-01,1.906838e-01,1.647210e-01,1.035789e-01,3.851359e-02,2.581207e-02,-1.540494e-02,-8.349884e-02,-1.273817e-01,-1.862502e-01,-2.414063e-01,-2.745722e-01,-3.473285e-01,-3.978962e-01,-4.152674e-01,-4.529329e-01,-4.819057e-01,-5.133201e-01,-5.919331e-01,-6.706371e-01,-7.269698e-01,-7.367399e-01,-7.499886e-01,-7.780852e-01,-8.161455e-01,-8.254012e-01,-8.721376e-01,-9.133787e-01,-9.225533e-01,-9.283349e-01,-9.263898e-01,-9.650131e-01,-9.850611e-01,-9.734718e-01,-9.900989e-01,-9.768719e-01,-9.754915e-01,-9.942748e-01,-9.782630e-01,-1.003032e+00,-9.615370e-01,-9.621956e-01,-9.373755e-01,-9.405639e-01,-9.061268e-01,-9.081392e-01,-8.966073e-01,-8.278345e-01,-8.612061e-01,-7.871536e-01,-7.676088e-01,-7.430801e-01,-6.823943e-01,-6.603251e-01,-6.314049e-01,-5.975769e-01,-5.434359e-01,-5.216082e-01,-4.625160e-01,-3.986903e-01,-3.599666e-01,-3.023848e-01,-2.258265e-01,-1.831760e-01,-1.501421e-01,-1.052318e-01,-5.672837e-02,-2.673988e-02,3.292604e-03,8.208407e-02,1.374082e-01,1.878373e-01,2.351748e-01,2.826925e-01,3.455181e-01,3.693189e-01,4.199062e-01,4.607233e-01,4.951458e-01,5.472112e-01,5.679768e-01,5.791349e-01,6.265096e-01,6.750147e-01,7.159085e-01,7.520997e-01,8.198636e-01,8.436612e-01,8.827533e-01,9.551504e-01,9.685745e-01,9.758965e-01,9.678210e-01,9.609784e-01,9.820225e-01,1.006042e+00,1.033114e+00,9.982639e-01,9.718402e-01,1.013435e+00,9.904421e-01,9.783145e-01,9.699275e-01,1.002672e+00,9.200357e-01,9.428122e-01,9.259162e-01,8.840212e-01,9.195512e-01,8.581934e-01,8.459096e-01,8.199683e-01,7.984706e-01,7.407096e-01,7.198053e-01,6.948672e-01,6.682850e-01,6.204460e-01,5.674909e-01,5.314585e-01,4.737505e-01,4.011681e-01,3.600472e-01,3.335377e-01,2.472971e-01,2.314435e-01,1.688678e-01,1.080565e-01,7.018775e-02,2.496548e-02,-2.268269e-02,-7.646536e-02,-1.205945e-01,-1.581059e-01,-1.982391e-01,-2.419851e-01,-3.020645e-01,-3.550158e-01,-4.244481e-01,-4.622000e-01,-5.070203e-01,-5.533374e-01,-6.033315e-01,-6.231499e-01,-6.476635e-01,-6.759727e-01,-7.164313e-01,-7.715893e-01,-7.972298e-01,-8.215141e-01,-8.665879e-01,-9.273024e-01,-9.383125e-01,-9.534628e-01,-9.407402e-01,-9.558069e-01,-9.713966e-01,-9.961733e-01,-9.829431e-01,-9.716396e-01,-1.005298e+00,-9.739617e-01,-1.014321e+00,-9.714029e-01,-9.898018e-01,-9.644133e-01,-9.899650e-01,-9.392489e-01,-9.484054e-01,-9.246187e-01,-9.211637e-01,-8.589989e-01,-8.748493e-01,-8.237362e-01,-8.020367e-01,-7.692074e-01,-7.618996e-01,-7.085815e-01,-6.519074e-01,-6.371682e-01,-5.612639e-01,-5.316684e-01,-4.689682e-01,-4.374304e-01,-4.015760e-01,-3.586355e-01,-3.202217e-01,-2.611625e-01,-2.160005e-01,-1.479536e-01,-1.030134e-01,-5.058643e-02,-3.081843e-02,4.508772e-02,1.238676e-01,1.694454e-01,1.784297e-01,2.364751e-01,2.821548e-01,3.089538e-01,3.643106e-01,3.958529e-01,4.208112e-01,4.650941e-01,5.252321e-01,5.602606e-01,6.120250e-01,6.685015e-01,7.263693e-01,7.600483e-01,7.980380e-01,8.203757e-01,8.161831e-01,8.466108e-01,8.950315e-01,9.099799e-01,9.170846e-01,9.459780e-01,9.308037e-01,9.678302e-01,9.986782e-01,9.972669e-01,1.003521e+00,1.020425e+00,9.938214e-01,9.830934e-01,9.836915e-01,9.951568e-01,1.013413e+00,9.540017e-01,9.521845e-01,9.321382e-01,8.958614e-01,8.571853e-01,8.880194e-01,8.065564e-01,8.273184e-01,7.786040e-01,7.475066e-01,7.171675e-01,6.515486e-01,6.529103e-01,6.214199e-01,5.635741e-01,4.861138e-01,4.571442e-01,4.082082e-01,3.789870e-01,3.509091e-01,2.883841e-01,2.265463e-01,1.797258e-01,1.163706e-01,7.266126e-02,1.565083e-02,-4.747275e-02,-7.456986e-02,-1.201368e-01,-1.637324e-01,-2.029170e-01,-2.786630e-01,-3.230430e-01,-3.528557e-01,-4.021535e-01,-4.673312e-01,-5.110804e-01,-5.710836e-01,-6.212538e-01,-6.504043e-01,-6.868987e-01,-7.106429e-01,-7.366216e-01,-7.624629e-01,-7.662139e-01,-8.093824e-01,-8.413509e-01,-8.604350e-01,-8.943303e-01,-9.137393e-01,-9.517299e-01,-9.618328e-01,-9.718617e-01,-1.017719e+00,-9.790716e-01,-9.934631e-01,-1.001183e+00,-1.009085e+00,-1.011207e+00,-1.030289e+00,-9.832156e-01,-9.704311e-01,-9.854128e-01,-9.266539e-01,-9.563838e-01,-9.126941e-01,-8.889444e-01,-8.552448e-01,-8.730401e-01,-8.208766e-01,-8.070195e-01,-7.307011e-01,-7.435801e-01,-6.708360e-01,-6.638763e-01,-6.042640e-01,-5.760834e-01,-5.473163e-01,-4.944649e-01,-4.708208e-01,-3.913187e-01,-3.599229e-01,-3.068135e-01,-2.522469e-01,-2.184183e-01,-1.647417e-01,-8.714265e-02,-2.615305e-02,5.237161e-03,2.652161e-02,5.790286e-02,9.668972e-02,1.804936e-01,2.651827e-01,3.037055e-01,3.447014e-01,3.673302e-01,4.047895e-01,4.631893e-01,4.878791e-01,5.111702e-01,5.522367e-01,5.774594e-01,6.139397e-01,6.657684e-01,7.289983e-01,7.929567e-01,8.271055e-01,8.472967e-01,8.410055e-01,8.523619e-01,8.807535e-01,9.669089e-01,9.679916e-01,9.735698e-01,9.702891e-01,9.738417e-01,1.015757e+00,1.011567e+00,9.682731e-01,9.925088e-01,9.968239e-01,9.694558e-01,9.881661e-01,9.613666e-01,1.003458e+00,9.155492e-01,9.510171e-01,8.562388e-01,8.940635e-01,8.239221e-01,8.488180e-01,8.110593e-01,7.939165e-01,7.715733e-01,6.994798e-01,6.732978e-01,6.520288e-01,5.955562e-01,5.501697e-01,5.134984e-01,4.489341e-01,4.126027e-01,3.668872e-01,3.058566e-01,2.650998e-01,2.300340e-01,1.761758e-01,8.425583e-02,5.310712e-02,3.301936e-02,-1.748670e-02,-7.092340e-02,-1.196330e-01,-1.482082e-01,-1.948236e-01,-2.520058e-01,-3.121484e-01,-3.594239e-01,-3.911695e-01,-4.496182e-01,-5.236944e-01,-5.656582e-01,-6.124573e-01,-6.583155e-01,-7.022225e-01,-7.290904e-01,-7.719639e-01,-7.899756e-01,-7.853036e-01,-8.019281e-01,-8.236061e-01,-8.236368e-01,-8.847394e-01,-9.139873e-01,-9.542297e-01,-9.597378e-01,-9.897676e-01,-9.938819e-01,-9.825987e-01,-1.003479e+00,-9.891601e-01,-1.001115e+00,-1.005976e+00,-1.007598e+00,-9.793871e-01,-9.993601e-01,-9.552637e-01,-9.386124e-01,-9.265302e-01,-9.354046e-01,-8.641562e-01,-8.442335e-01,-8.448187e-01,-8.494816e-01,-7.633925e-01,-7.491192e-01,-7.458119e-01,-6.634827e-01,-6.345058e-01,-6.183563e-01,-5.743961e-01,-5.469036e-01,-5.213535e-01,-4.687420e-01,-4.054345e-01,-3.632679e-01,-2.955694e-01,-2.271318e-01,-1.651211e-01,-1.332020e-01,-9.570618e-02,-8.832947e-03,1.440568e-02};
 
-        NLMSFilter filt1 = new NLMSFilter(mu, length, NLMSFilter.WeightsFillMethod.ZEROS);
+        LMS filt1 = new LMS(mu, length, LMS.WeightsFillMethod.ZEROS);
         filt1.filter(this.desired2, this.signal2);
-        NLMSFilter filt2 = new NLMSFilter(mu, weights);
+        LMS filt2 = new LMS(mu, weights);
         filt2.filter(this.desired2, this.signal2);
 
-        Assertions.assertArrayEquals(result, filt1.getOutput(), 0.1);
-        Assertions.assertArrayEquals(result, filt2.getOutput(), 0.1);
+        Assertions.assertArrayEquals(result, filt1.getOutput(), 0.05);
+        Assertions.assertArrayEquals(result, filt2.getOutput(), 0.05);
 
         // Save plot result
         Plotting fig = new Plotting(600, 500, "Sample Figure", "Time", "Signal");
